refactor: table APIs to connection_view.py
